third_party/WebKit/LayoutTests/fast/forms/text/text-select-disabled.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<input id="visible" value="text0">
<input id="disabled" value="text1" disabled>
<script>
description("This test confirms that select an disabled input element does not cause focus change.");
visible.select();
shouldBe('getSelection().anchorNode', 'document.body');
shouldBe('getSelection().anchorOffset', '3');
disabled.select();
shouldBe('getSelection().anchorNode', 'document.body');
shouldBe('getSelection().anchorOffset', '3');
</script>
</body>
</html>
