third_party/blink/web_tests/fast/forms/text/text-select-disabled-expected.txt

This test confirms that select an disabled input element does not cause focus change.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getSelection().anchorNode is document.body
PASS getSelection().anchorOffset is 3
PASS getSelection().anchorNode is document.body
PASS getSelection().anchorOffset is 3
PASS successfullyParsed is true

TEST COMPLETE

 
